Added save_to_hdf to PSA.
